refactor: move execution info to a separate file
