+++ app/packages/core/src/plugins/SchemaIO/components/NativeModelEvaluationView/components/ColorSquare.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+export default function ColorSquare(props: { color: string }) {
+  return (
+    <div
+      style={{
+        width: 16,
+        height: 16,
+        backgroundColor: props.color,
+        borderRadius: 2,
+      }}
+    />
+  );
+}
